Separate package for database connection
